登录功能

+++ src/main/java/org/example/controller/page/UserPageController.java
@@ -9,6 +9,9 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+
+import javax.servlet.http.HttpSession;
+import java.util.List;
 import java.util.Objects;
 
 @Controller
@@ -16,21 +19,39 @@ public class UserPageController {
 
     @Autowired
     private UserService userService;
+    @Autowired
+    private HttpSession session;
+
+    @GetMapping("userList")
+    public String userList(Model model) {
+        UserDO userDO = (UserDO) session.getAttribute("user");
+        if (Objects.isNull(userDO)) {
+            return "redirect:/";
+        } else {
+            List<UserDO> users = (List<UserDO>) userService.findAll();
+            model.addAttribute("users", users);
+            return "user/userList";
+        }
+    }
 
     @GetMapping("userAdd")
-    @NoResponseBody
     public String userAdd() {
         return "user/userUpdate";
     }
 
     @GetMapping("userUpdate/{id}")
     public String userUpdate(@PathVariable Long id, Model model) {
-        model.addAttribute("user",userService.findById(id));
-        return "user/userUpdate";
+        UserDO userDO = (UserDO) session.getAttribute("user");
+        if (Objects.isNull(userDO)) {
+            return "redirect:/";
+        } else {
+            model.addAttribute("user",userService.findById(id));
+            return "user/userUpdate";
+        }
     }
 
     @PostMapping("userSave")
-    public String save(UserDO user, Model model) {
+    public String save(UserDO user) {
         if (Objects.nonNull(user)) {
             if (Objects.nonNull(user.getId())) {
                 userService.update(user);
@@ -38,12 +59,15 @@ public class UserPageController {
                 userService.add(user);
             }
         }
-        return "redirect:/";
+        return "redirect:userList";
     }
 
     @GetMapping("userDelete/{id}")
     public String delete(@PathVariable Long id) {
-        userService.deleteById(id);
-        return "redirect:/";
+        UserDO userDO = (UserDO) session.getAttribute("user");
+        if (!id.equals(userDO.getId())) {
+            userService.deleteById(id);
+        }
+        return "forward:userList";
     }
 }
